refactor: 패키지 수정에 따른 오류 해결
