+++ src/test/java/cart/application/PaymentServiceTest.java
@@ -7,8 +7,8 @@ import cart.domain.PurchaseOrderInfo;
 import cart.domain.PurchaseOrderItem;
 import cart.domain.point.Point;
 import cart.domain.point.UsedPoint;
-import cart.dto.purchaseorder.PurchaseOrderItemRequest;
-import cart.dto.purchaseorder.PurchaseOrderRequest;
+import cart.dto.purchaseorder.request.PurchaseOrderItemRequest;
+import cart.dto.purchaseorder.request.PurchaseOrderRequest;
 import cart.exception.PointException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
